data-collection/requirements.txt
# Requirements for Cryptocurrency Data Crawler

# Core libraries
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0

# Data storage
openpyxl>=3.1.0

# Scheduling (optional)
schedule>=1.2.0

# Progress bar (optional)
tqdm>=4.65.0

# API rate limiting
ratelimit>=2.2.1

# Beautiful output
colorama>=0.4.6
tabulate>=0.9.0


s3fs==2024.6.1
pyarrow==21.0.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
kafka-python==2.2.15
